Add option to use (v, f) signature on callbacks
This boilerplate depends on trimesh but I wasn't actually using trimesh
in other projects. So for flexibility allow vertices and faces to be the input
of the main geometry processing callback.

Right now it does so by calling trimesh object's vertices and faces but this
is rather fragile, what if they change mesh.vertices data type?